faster test using more idiomatic function calls for standalone integration tests
