+++ itests/http_tests.rs
@@ -1,0 +1,7 @@
+// Test harness for all integration tests
+mod common;
+mod test;
+
+mod oxen_repo_committed_files_should_return_real_data_via_http_get;
+mod oxen_repo_held_csv_should_be_accessible_via_http_get;
+mod oxen_server_health_should_be_accessible_via_http_get;
