--- tests/oxen_repo_held_csv_should_be_accessible_via_http_get.rs
@@ -1,148 +1,0 @@
-use std::process::{Child, Command, Stdio};
-use std::time::Duration;
-use tokio::time::sleep;
-
-struct TestServer {
-    child: Child,
-    base_url: String,
-}
-
-impl TestServer {
-    /// Start a real oxen-server process for integration testing
-    async fn start() -> Result<Self, Box<dyn std::error::Error>> {
-        let sync_dir = liboxen::test::test_run_dir().join("integration_http_test");
-        Self::start_with_sync_dir(&sync_dir).await
-    }
-    
-    /// Start a real oxen-server process with custom sync directory
-    async fn start_with_sync_dir(sync_dir: &std::path::Path) -> Result<Self, Box<dyn std::error::Error>> {
-        // Create the sync directory
-        std::fs::create_dir_all(&sync_dir)?;
-        
-        // Find the oxen-server binary
-        let server_path = std::env::current_dir()?
-            .join("target")
-            .join("debug")
-            .join("oxen-server");
-            
-        if !server_path.exists() {
-            return Err("oxen-server binary not found. Run 'cargo build' first".into());
-        }
-        
-        // Start the server process
-        let mut child = Command::new(server_path)
-            .arg("start")
-            .arg("--ip")
-            .arg("127.0.0.1")
-            .arg("--port")
-            .arg("3001") // Use different port to avoid conflicts
-            .env("SYNC_DIR", &sync_dir)
-            .stdout(Stdio::null()) // Suppress output to avoid hanging
-            .stderr(Stdio::null())
-            .spawn()?;
-            
-        // Wait for server to start
-        sleep(Duration::from_secs(2)).await;
-        
-        // Check if process is still running
-        match child.try_wait() {
-            Ok(Some(status)) => {
-                return Err(format!("Server process exited early with status: {}", status).into());
-            }
-            Ok(None) => {
-                // Process is still running, good
-            }
-            Err(e) => {
-                return Err(format!("Error checking server process: {}", e).into());
-            }
-        }
-        
-        // Try to connect to health endpoint to verify server is ready
-        let client = reqwest::Client::new();
-        let base_url = "http://127.0.0.1:3001".to_string();
-        
-        for _ in 0..10 {
-            if let Ok(response) = client.get(&format!("{}/api/health", base_url)).send().await {
-                if response.status().is_success() {
-                    return Ok(TestServer {
-                        child,
-                        base_url,
-                    });
-                }
-            }
-            sleep(Duration::from_millis(500)).await;
-        }
-        
-        // If we get here, server didn't start properly
-        let _ = child.kill();
-        Err("Server failed to start or health check failed".into())
-    }
-    
-    /// Get the base URL for this test server
-    fn base_url(&self) -> &str {
-        &self.base_url
-    }
-}
-
-impl Drop for TestServer {
-    fn drop(&mut self) {
-        // Clean up the server process
-        let _ = self.child.kill();
-        let _ = self.child.wait();
-    }
-}
-
-#[tokio::test]
-async fn oxen_repo_held_csv_should_be_accessible_via_http_get() {
-    // This test focuses specifically on CSV file accessibility via HTTP GET
-    // Create a test repository with CSV data
-    let test_dir = std::env::temp_dir().join("oxen_csv_test");
-    let _ = std::fs::remove_dir_all(&test_dir);
-    std::fs::create_dir_all(&test_dir).expect("Failed to create test directory");
-    
-    // Create repository with CSV file
-    let repo_dir = test_dir.join("test_user").join("csv_repo");
-    std::fs::create_dir_all(&repo_dir).expect("Failed to create repo directory");
-    
-    let csv_content = "product,price,category\nLaptop,999.99,Electronics\nChair,149.50,Furniture\nBook,19.99,Education";
-    std::fs::write(repo_dir.join("products.csv"), csv_content).expect("Failed to write CSV");
-    
-    // Initialize Oxen repo and commit CSV
-    let oxen_binary = std::env::current_dir().unwrap().join("target/debug/oxen");
-    if !oxen_binary.exists() {
-        panic!("oxen binary not found. Run 'cargo build --bin oxen' first");
-    }
-    
-    // Oxen VCS workflow: init -> add -> commit
-    std::process::Command::new(&oxen_binary).arg("init").current_dir(&repo_dir).output().expect("Failed to init");
-    std::process::Command::new(&oxen_binary).args(&["config", "--name", "Test", "--email", "test@test.com"]).current_dir(&repo_dir).output().unwrap();
-    std::process::Command::new(&oxen_binary).args(&["add", "products.csv"]).current_dir(&repo_dir).output().expect("Failed to add CSV");
-    std::process::Command::new(&oxen_binary).args(&["commit", "-m", "Add CSV data"]).current_dir(&repo_dir).output().expect("Failed to commit");
-    
-    // Start oxen-server
-    let server = TestServer::start_with_sync_dir(&test_dir).await.expect("Failed to start test server");
-    
-    // Create HTTP client
-    let client = reqwest::Client::new();
-    
-    // Test: HTTP GET should return the CSV file content
-    let response = client
-        .get(&format!("{}/api/repos/test_user/csv_repo/files/main/products.csv", server.base_url()))
-        .send()
-        .await
-        .expect("Failed to send HTTP GET request for CSV");
-    
-    let status = response.status();
-    println!("CSV HTTP GET response status: {}", status);
-    let body = response.text().await.expect("Failed to read CSV response body");
-    println!("CSV HTTP GET response body: {}", body);
-    
-    // Verify we can access CSV data via HTTP GET
-    if status.is_success() && body.contains("Laptop,999.99,Electronics") {
-        println!("✅ CSV file successfully accessible via HTTP GET!");
-    } else {
-        println!("⚠️  CSV file not accessible via HTTP GET - status: {}", status);
-        // This is expected to fail until the full CSV endpoint is implemented
-    }
-}
-
